requirements.txt
alembic==1.5.2
click==7.1.2
contourpy==1.0.7
cycler==0.11.0
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
fonttools==4.39.2
gunicorn==20.0.4
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.4.4
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.7.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
psycopg2-binary==2.9.5
pyparsing==3.0.9
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2022.7.1
seaborn==0.12.2
six==1.15.0
SQLAlchemy==1.3.22
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
